vmd-script, random seeds, rewriting
I added a vmd scripts to visualize the simulation. Many users might be
interested in that, I think.
I tried to improve reproducibility of the simulations, to enable the
detailed discussion of the trajectories, but in the end it is still not
exactly reproducible. Therefore, I rewrote parts of the discussion,
since the precise mentioning of the certain frames in the simulations
is not meaningful...