revise tutorial (#6)
* vmd-script, random seeds, rewriting

I added a vmd scripts to visualize the simulation. Many users might be
interested in that, I think.
I tried to improve reproducibility of the simulations, to enable the
detailed discussion of the trajectories, but in the end it is still not
exactly reproducible. Therefore, I rewrote parts of the discussion,
since the precise mentioning of the certain frames in the simulations
is not meaningful...

* minor change: change TOC headline

rewrote the bullet point for the visualization of vmd

* simplify simulation.mdp

reverted changes in this file to make it more simple again.
In comparison to main branch, the only difference are the fixed random
seeds.

* fewer steps in tutorual sim.

* remove installation from tutorial

We removed the explicit installation from the tutorial and refer to the
README for installation instructions instead.

* add example figures

I generated figures from the trajectory in the output-worked path.
added those there. In upcoming commits, These should be included in the tutorial notebook for
reference.

* Add example images to tutorial notebook

* comment on special case: chain permutation

It actually happened to me now once that the structure assembled in a
different permutation as in the crystal structure. Therefore, the RMSD
of the complex never approached zero. I added a few sentences in the
tutorial notebook to explain this eventual case.

* remove cell-ouputs...

* remove typo

* add figs for permutated assembly

also added the trajectory with the alternative assembly in the
output-worked path.

* Add comment for permutation invariant RMSD calculation

---------

Co-authored-by: Patrick Quoika <patrick.quoika@tum.de>
Co-authored-by: Luis Walter <luis.walter@tum.de>